requirements.txt
certifi==2016.8.8
future==0.15.2
peewee==2.8.3
python-telegram-bot==5.0.0
requests==2.11.1
urllib3==1.16
